Use if bindings for Invite status display
This might seem like working backwards, but I feel it's necessary to be able to understand (to a higher degree) the logic happening inside html pages. This commit adds information back to the html Invite page that makes it easier to update and understand months down the road when its no longer a relevant point of interest.

Magic is helpful, but sometimes it can cause unnecessary complexity and ambiguity.