Add validateAlias() client API method
Fix client API validateInvite() to use proper responses

The server currently responds with an unnecessary object instead of a proper response, to the query: "Is this invite valid?". The API has been updated to refect how it will respond in the future and how it SHOULD have been designed from the beginning.