Update(app): add scroll reset on route change
Whenever the content changes inside the body, it will now
reset the scroll position to the top, which is expected behavior
when navigating to a different page.

The caveat however is that there might be times when I
need to remember a scroll position, but I can handle those
on a case-by-case basis, within that specific view.